% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{createCohortsFun}
\alias{createCohortsFun}
\title{Load default content for createCohorts.R}
\usage{
createCohortsFun(path = ".")
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default \code{"."}.}
}
\value{
A character vector containing the default contents of \code{createCohorts.R}.
}
\description{
Load default content for createCohorts.R
}
