% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogger.R
\name{createLogger}
\alias{createLogger}
\title{`createLogger()` as a text file in a results directory}
\usage{
createLogger(resultsDir)
}
\arguments{
\item{resultsDir}{A valid folder where to save the results of a study.
This function will work best with the folder structure formed by
`createsResultsDir()`}
}
\value{
Invisible
}
\description{
Sets 'ohdsi/ParallelLogger' log and error report for a study
}
\examples{
directories <- createResultsDir()
createLogger(
  resultsDir = directories$resultsDir
  )

}
