% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{developCheckout}
\alias{developCheckout}
\title{`developCheckout()` is a wrapper for gert functions to swiftly default to develop
and pull latest changes}
\usage{
developCheckout()
}
\value{
Git log messages after checking out and pulling 'develop'
}
\description{
`developCheckout()` is a wrapper for gert functions to swiftly default to develop
and pull latest changes
}
