% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertDocs}
\alias{insertDocs}
\title{Insert default documentation files}
\usage{
insertDocs(path)
}
\arguments{
\item{path}{A valid path to the study project in character.

Creates default \code{NEWS.md}, \code{README.Rmd} and \code{LICENSE.md} (Apache License 2.0) files.}
}
\value{
No return value.
}
\description{
Insert default documentation files
}
