% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertTests}
\alias{insertTests}
\title{Insert default test files}
\usage{
insertTests(path = ".")
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default \code{"."}.}
}
\value{
No return value.
}
\description{
Creates the \code{tests} folder and populates the \code{tests/testthat} folder with
a default test file for each script found in the \code{R} folder.
}
