% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCellar.R
\name{installCellar}
\alias{installCellar}
\title{\code{installCellar()} install files already saved in a cellar}
\usage{
installCellar(path, project, library, lockfile)
}
\arguments{
\item{path}{Character string giving the cellar path. If omitted, defaults to \code{renv/cellar}.}

\item{project}{Character string giving the project path. If omitted, defaults to \code{usethis::proj_get()}.}

\item{library}{Character string giving the library path. If omitted, defaults to \code{renv/library}.}

\item{lockfile}{Character string giving the lockfile path. If omitted, defaults to \code{renv.lock} at the project base level provided by \code{usethis::proj_get()}.}
}
\value{
Invisible
}
\description{
\code{installCellar()} install files already saved in a cellar
}
\examples{

\dontrun{
# It will install all packages in the cellar
installCellar()
}
}
