% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{issueOpen}
\alias{issueOpen}
\title{\code{issueOpen()} is a wrapper for gh::gh() to quickly create an issue
in the current GitHub repository.}
\usage{
issueOpen(title, body, newBranch = FALSE)
}
\arguments{
\item{title}{A character string giving the issue title.}

\item{body}{A character string giving the issue body.}

\item{newBranch}{Logical. Whether to create and check out a branch named
after the new issue. Defaults to \code{FALSE}.}
}
\value{
URL of the created issue.
}
\description{
Set GITHUB_PAT in .Renviron. It will use the same authentication
in the background as with \code{gh} https://gh.r-lib.org/articles/managing-personal-access-tokens.html
}
\examples{
\dontrun{
# It requires GITHUB_PAT
issueOpen(
   title = "Add study outcome",
   body = "Describe the planned change.",
   newBranch = TRUE
)
}
}
