% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{runDiagnosticsFun}
\alias{runDiagnosticsFun}
\title{Load default content for runDiagnostics.R}
\usage{
runDiagnosticsFun(path = ".")
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default \code{"."}.}
}
\value{
A character vector containing the default contents of \code{runDiagnostics.R}.
}
\description{
Load default content for runDiagnostics.R
}
