% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{runStudyFun}
\alias{runStudyFun}
\title{Load default content for runStudy.R}
\usage{
runStudyFun(n_obj = 3)
}
\arguments{
\item{n_obj}{Number of study objectives to include in the workflow,
default \code{3}.}
}
\value{
A character vector containing the default contents of \code{runStudy.R}.
}
\description{
Load default content for runStudy.R
}
