% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLoggers.R
\name{setLoggers}
\alias{setLoggers}
\title{Configure study loggers that write text files to the results directory}
\usage{
setLoggers(
  resultsDir,
  logFileName = "log",
  errorFileName = "error",
  eventLevel = "TRACE",
  errorLevel = "ERROR"
)
}
\arguments{
\item{resultsDir}{A valid folder where to save the results of a study.
This function will work best with the folder structure formed by
\code{createResultsDir()}}

\item{logFileName}{A character string specifying the file name. Default 'log'}

\item{errorFileName}{A character string specifying the file name. Default 'error'}

\item{eventLevel}{TRACE is the default, captures all the output from
the console}

\item{errorLevel}{ERROR is the default, captures errors and fatal events}
}
\value{
Invisible
}
\description{
Configures the \code{ohdsi/ParallelLogger} logger and error report for a study. The logger
is registered as 'OMOP_STUDY_LOGGER' and the error report as 'OMOP_STUDY_ERROR_REPORT'.
Please note: avoid using it inside \code{tryCatch()} or similar because events
will not be observed or recorded by ParallelLogger.
}
\examples{
outputDir <- file.path(
   tempdir(),
   "examples"
)
directories <- createResultsDir(
   outputDir,
   dbname = "OMOP"
   )
setLoggers(
  resultsDir = directories$resultsDir
  )
ParallelLogger::clearLoggers()
unlink(outputDir, recursive = TRUE)
}
