% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipStudyFiles.R
\name{zipStudyFiles}
\alias{zipStudyFiles}
\title{\code{zipStudyFiles()} compresses study results}
\usage{
zipStudyFiles(outputDir, resultsDirName, dbname)
}
\arguments{
\item{outputDir}{Character string giving the path to the base output directory.}

\item{resultsDirName}{Character string giving the name or path of the results subfolder.}

\item{dbname}{Character string giving the database identifier.}
}
\value{
Character string containing the file name of the generated zip archive.
}
\description{
\code{zipStudyFiles()} compresses study results
}
\examples{
outputDir <- file.path(
   tempdir(),
   "examples"
   )
dbname <- "IPCI"
directories <- createResultsDir(
  outputDir,
  dbname = dbname
)
zipStudyFiles(
  resultsDirName = directories$resultsDir,
  outputDir = directories$outputDir,
  dbname = dbname
)
unlink(outputDir, recursive = TRUE)
}
