% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmNames.R
\name{arrangeCdmNames}
\alias{arrangeCdmNames}
\title{`arrangeCdmNames()` verifies if data partner's acronyms are valie
and returns a vector ordered by country alphabetical order}
\usage{
arrangeCdmNames(labels)
}
\arguments{
\item{labels}{A character vector of data partners acronyms}
}
\value{
A character vector
}
\description{
`arrangeCdmNames()` verifies if data partner's acronyms are valie
and returns a vector ordered by country alphabetical order
}
\examples{
# Verifies acronyms are valid and sorts them in the correct order
labels <- c(
  "BCR",
  "IQVIA US - PMTX+",
  "IQVIA US - AmbEMR",
  "IQVIA LPD Belgium",
)
arrangeCdmNames(labels = labels)
}
