% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangeCdmName.R
\name{arrangeCdmNames}
\alias{arrangeCdmNames}
\title{`arrangeCdmNames()` retrieves data partners acronyms arranged by country alphabetical order}
\usage{
arrangeCdmNames(labels)
}
\arguments{
\item{labels}{A character vector to filter specific cdm_names}
}
\value{
A character vector
}
\description{
`arrangeCdmNames()` retrieves data partners acronyms arranged by country alphabetical order
}
\examples{
# Filter a group of acryonims
labels <- c(
  "BCR",
  "IQVIA LPD Belgium",
  "IQVIA US - AmbEMR",
  "IQVIA US - PMTX+"
)
arrangeCdmNames(labels = labels)
}
