% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmNames.R
\name{assertCdmNames}
\alias{assertCdmNames}
\title{`assertCdmNames()` verifies if data partner's names are correct and display an error if incorrect}
\usage{
assertCdmNames(labels, expected)
}
\arguments{
\item{labels}{A character vector of data partner acronyms}

\item{expected}{A character vector of expected data partner acronyms. If omitted, the package's built-in CDM names are used.}
}
\value{
Invisible if labels are correct
}
\description{
`assertCdmNames()` verifies if data partner's names are correct and display an error if incorrect
}
\examples{
# Assert if acronyms are correct
labels <- c(
  "BCR",
  "IQVIA LPD Belgium",
  "IQVIA US - AmbEMR",
  "IQVIA US - PMTX+"
)
assertCdmNames(labels = labels)
}
