% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLoggers.R
\name{createLoggers}
\alias{createLoggers}
\title{`createLoggers()` as a text file in a results directory}
\usage{
createLoggers(
  resultsDir,
  loggerName = "OMOP STUDY",
  logFileName = "log.txt",
  errorFileName = "error.txt",
  eventLevel = "TRACE"
)
}
\arguments{
\item{resultsDir}{A valid folder where to save the results of a study.
This function will work best with the folder structure formed by
`createsResultsDir()`}

\item{eventLevel}{TRACE is the default, captures all the output from
the console}
}
\value{
Invisible
}
\description{
Sets `ohdsi/ParallelLogger` log and error report for a study
}
\examples{
directories <- createResultsDir()
createLogger(
  resultsDir = directories$resultsDir
  )
}
