% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsDir.R
\name{createResultsDir}
\alias{createResultsDir}
\title{Create directory for analysis results}
\usage{
createResultsDir(outputDir = NULL, dbname)
}
\arguments{
\item{outputDir}{Character string. The path to the main output folder.
If NULL, defaults to the current working directory. Default NULL.}

\item{dbname}{Character string. The name of the database, used to suffix
the results folder (e.g., "results_dbname").}
}
\value{
A list containing three elements:
\itemize{
\item \code{outputDir}: The path to the main output directory.
\item \code{resultsDir}: The path to the specific results subdirectory created.
\item \code{resultsDirName}: The name of the results subdirectory.
}
}
\description{
Creates a subdirectory for results based on the database name within a
specified output directory. If the output directory is not provided, the
current working directory is used.
}
\examples{
outputDir <- file.path(
   tempdir(),
   "examples"
)
directories <- createResultsDir(
   outputDir,
   dbname = "OMOP"
   )
unlink(outputDir, recursive = TRUE)
}
