% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{devCheckout}
\alias{devCheckout}
\title{\code{devCheckout()} checks out the \code{develop} branch and pulls its latest changes.}
\usage{
devCheckout()
}
\value{
Git log messages after checking out and pulling 'develop'
}
\description{
Checks out the \code{develop} branch and pulls changes from its remote.
}
\examples{
\dontrun{
# Requires access to the configured Git remote
devCheckout(
)
}
}
