% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{devCheckout}
\alias{devCheckout}
\title{\code{devCheckout()} is a wrapper for gert functions to swiftly default to develop
and pull latest changes}
\usage{
devCheckout()
}
\value{
Git log messages after checking out and pulling 'develop'
}
\description{
Set GITHUB_PAT in .Renviron. It will use the same authentication
in the background as with \code{gh} https://gh.r-lib.org/articles/managing-personal-access-tokens.html
}
\examples{
\dontrun{
# It requires GITHUB_PAT
devCheckout(
)
}
}
