% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPkgZips.R
\name{getPkgZips}
\alias{getPkgZips}
\title{Download zip-compressed Windows binaries from a renv.lock file}
\usage{
getPkgZips(
  lockfile_path = NULL,
  supplement = NULL,
  override_lock = FALSE,
  r_rels_vect = c("4.0", "4.1", "4.2", "4.3", "4.4", "4.5", "4.6"),
  backupRrel = "4.4",
  outputDir
)
}
\arguments{
\item{lockfile_path}{Path to renv.lock file}

\item{supplement}{Nested list of packages to supplement or use in place of a renv.lock file}

\item{override_lock}{Logical. Default is FALSE. If set to TRUE and the same package is named in both the renv.lock and the supplement, the version called in the supplement will override the renv.lock}

\item{r_rels_vect}{A character vector of R minor releases}

\item{backupRrel}{A string of the R minor-release to use if preferred package version cannot be found under any of the minor releases specified in `r_rels_vect`}

\item{outputDir}{Path to output directory for package zips}
}
\value{
In addition to downloading the specified packages into the outputDir, returns a list of packages that were searched for as well as the URLs, versions (if found), etc.
}
\description{
A function that locates and downloads Windows binaries (.zip) from CRAN based on packages specified in a renv.lock file.
}
\details{
If the package version is not found under any of the R minor releases, the package will be downloaded from the specified `backupRrel` regardless of package version specified in renv.lock
}
\examples{
pkg_status_list <- getPkgZips()

# Format supp in this way:
devtools <- list(Package = "devtools", Version = "2.5.1")
duckdb <- list(Package = "duckdb", Version = "1.5.2")
supp <- list(Packages = list(devtools = devtools, duckdb = duckdb))

pkg_status_list <- getPkgZips(supplement = supp)

}
