% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPkgZips.R
\name{getPkgZips}
\alias{getPkgZips}
\title{Download zip-compressed Windows binaries from a renv.lock file}
\usage{
getPkgZips(
  lockfile_path,
  r_rels_vect = c("4.0", "4.1", "4.2", "4.3", "4.4", "4.5", "4.6"),
  backupRrel = "4.4",
  cellarDir
)
}
\arguments{
\item{lockfile_path}{Path to renv.lock file}

\item{r_rels_vect}{A character vector of R minor releases}

\item{backupRrel}{A string of the R minor-release to use if preferred package version cannot be found under any of the minor releases specified in `r_rels_vect`}

\item{cellarDir}{Path to renv-cellar}
}
\value{
Returns a list of packages that were searched for as well as the URLs, versions (if found), etc.
}
\description{
A function that locates and downloads Windows binaries (.zip) from CRAN based on packages specified in a renv.lock file.
}
\details{
In the case that the package version is not found under any of the R minor releases, the package will be downloaded from the specified `backupRrel` regardless of package version specified in renv.lock
}
\examples{
\dontrun{
pkg_status_list <- getPkgZips(lockfile_path = "path/to/renv.lock", r_rels_vect = c("4.3", "4.4", "4.5", "4.6"), backupRrel = "4.5", cellarDir = "path/to/renv-cellar")
}

}
