% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertDocs}
\alias{insertDocs}
\title{Insert default documentation files}
\usage{
insertDocs(path)
}
\value{
No return value.
}
\description{
Creates default \code{NEWS.md}, \code{README.Rmd} and \code{LICENSE.md} (Apache License 2.0) files.
}
