% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertStructure}
\alias{insertStructure}
\title{Set up study structure}
\usage{
insertStructure(path = ".", n_obj = 3)
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default \code{"."}.}

\item{n_obj}{Number of study objectives, default \code{n_obj = 3}.}
}
\value{
No return value.
}
\description{
Creates the standard structure required for a new OMOP study package.
It includes standard folders, R scripts, test and documentation files and it
installs a default list of DARWIN EU®/OHDSI packages.
}
\details{
It assumes an active R project has already been created and the \code{renv}
initialized.
}
