% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertStudyFiles}
\alias{insertStudyFiles}
\title{Insert default study files}
\usage{
insertStudyFiles(path = ".", n_obj = 3)
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default \code{"."}.}

\item{n_obj}{Number of study objectives, default \code{n_obj = 3}.}
}
\value{
No return value.
}
\description{
Creates the \code{inst} and \code{extras} folders and populates the \code{R} folder with
ready files for standard study functions.
}
