% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{insertTests}
\alias{insertTests}
\title{Insert default test files}
\usage{
insertTests(path = ".")
}
\arguments{
\item{path}{Character string identifying the path to the study project,
default `"."`.}
}
\value{
No return value.
}
\description{
Creates the `tests` folder and populates the `tests/testthat` folder with
a default test file for each script found in the `R` folder.
}
