% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCellar.R
\name{installCellar}
\alias{installCellar}
\title{\code{installCellar()} install files already saved in a cellar}
\usage{
installCellar(path, project, library, lockfile)
}
\arguments{
\item{path}{Cellar path in character. If missing defaults to 'renv/cellar'.}

\item{project}{Project path in character. If missing defaults to \code{usethis::proj_get()}.}

\item{library}{Library path in characcter. If missing defauls to 'renv/library'.}

\item{lockfile}{Lockfile file path in character. If missing defaults to 'renv.lock' at the project base level provided by \code{usethis::proj_get()}.}
}
\value{
Invisible
}
\description{
\code{installCellar()} install files already saved in a cellar
}
\examples{

\dontrun{
# It will install all packages in the cellar
installCellar()
}
}
