% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertStructure.R
\name{installPackageBundle}
\alias{installPackageBundle}
\title{Install default package bundle}
\usage{
installPackageBundle(path)
}
\arguments{
\item{path}{A valid path to the study project in character.

Installs a predefined set of DARWIN EU®/OHDSI packages into the study project
and updates their dependencies. The standard package list includes the packages
omopgenerics, PhenotypeR, DrugExposureDiagnostics, CohortConstructor,
IncidencePrevalence, DrugUtilisation, CohortCharacteristics, CohortSurvival,
visOmopResults and DarwinShinyModules.}
}
\value{
An invisible list of the installed packages.
}
\description{
Install default package bundle
}
