% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{issueOpen}
\alias{issueOpen}
\title{\code{issueOpen()} is a wrapper for gh::gh() to swiftly post
in the GitHub repository of the current project.}
\usage{
issueOpen(title, body, newBranch = FALSE)
}
\arguments{
\item{title}{of the issue in character}

\item{body}{of the issue in character}

\item{newBranch}{Logical. Default TRUE, will open a new
branch in GitHub format}
}
\value{
A message with the link of the PR
}
\description{
Set GITHUB_PAT in .Renviron. It will use the same authentication
in the background as with \code{gh} https://gh.r-lib.org/articles/managing-personal-access-tokens.html
}
\examples{
\dontrun{
# It requires GITHUB_PAT
issueOpen(
   title = "An issue title",
   body = "Description of issue",
   newBranch = FALSE
)
}
}
