% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCellar.R
\name{packageCellar}
\alias{packageCellar}
\title{\code{packageCellar()} saves files from a renv.lock file into the 'cellar' folder in renv}
\usage{
packageCellar(path, project, lockfile, type = "complete")
}
\arguments{
\item{path}{Cellar path in character. If missing defaults to 'renv/cellar'.}

\item{project}{Project path in character. If missing defaults to \code{usethis::proj_get()}.}

\item{lockfile}{Lockfile file path in character. If missing defaults to 'renv.lock' at the project base level provided by \code{usethis::proj_get()}.}

\item{type}{A choice in characcter from either "complete" or "github" to download only packages from GH.}
}
\value{
Invisible
}
\description{
\code{packageCellar()} saves files from a renv.lock file into the 'cellar' folder in renv
}
\keyword{internal}
