% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCellar.R
\name{packageCellar}
\alias{packageCellar}
\title{\code{packageCellar()} saves files from a renv.lock file into the 'cellar' folder in renv}
\usage{
packageCellar(lockfile, cellarDir, type = "complete")
}
\arguments{
\item{lockfile}{A valid path to a lockfile in character.}

\item{cellarDir}{A valid path to the cellar in character. If no value is assigned, it will create renv/cellar}

\item{type}{A choice in characcter from either "complete" or "github" to download only packages from GH.}
}
\value{
Invisible
}
\description{
\code{packageCellar()} saves files from a renv.lock file into the 'cellar' folder in renv
}
\keyword{internal}
