% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{pullRequest}
\alias{pullRequest}
\title{\code{pullRequest()} is a wrapper for gh::gh() to quickly create a pull request
to merge the current branch.}
\usage{
pullRequest(title, body, base = "develop")
}
\arguments{
\item{title}{A character string giving the pull request title.}

\item{body}{A character string giving the pull request body.}

\item{base}{A character string giving the target branch. Defaults to "develop".}
}
\value{
URL of the created pull request.
}
\description{
Set GITHUB_PAT in .Renviron. It will use the same authentication
in the background as with \code{gh} https://gh.r-lib.org/articles/managing-personal-access-tokens.html
}
\examples{
\dontrun{
# It requires GITHUB_PAT
pullRequest(
   title = "A PR title",
   body = "Description of PR",
   base = "develop"
)
}
}
