% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionControl.R
\name{pullRequest}
\alias{pullRequest}
\title{\code{pullRequest()} is a wrapper for gh::gh() to swiftly ask merging
code from the current branch}
\usage{
pullRequest(title, body, base = "develop")
}
\arguments{
\item{title}{of the issue in character}

\item{body}{of the issue in character}

\item{base}{The target branch, in character. Defaults to "develop"}
}
\value{
A message with the link of the issue
}
\description{
\code{pullRequest()} is a wrapper for gh::gh() to swiftly ask merging
code from the current branch
}
