% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipStudyFiles.R
\name{unZipStudyFiles}
\alias{unZipStudyFiles}
\title{Unzip study result files}
\usage{
unZipStudyFiles(path, recursive = FALSE, pattern, negate = TRUE, outputDir)
}
\arguments{
\item{path}{Character directory where zip files are searched.}

\item{recursive}{Logical indicating whether to search recursively for zip files.}

\item{pattern}{Optional regular expression used to filter zip file paths.
For instance, just unzip those files that do not belong to a 'DED' folder.}

\item{negate}{Logical passed to \code{stringr::str_detect()} to negate the optional pattern argument.}

\item{outputDir}{Character directory where files will be unzipped.}
}
\value{
Invisible called for its side effects.
}
\description{
Finds zip files under \code{path} and extracts them to \code{outputDir}.
}
\examples{
 \dontrun{
# It will unzip folders in temp file
path <- testthat::test_path(
   "data",
   "results_execution"
)
outputDir <- file.path(
  tempdir(),
  "examples"
)
unZipStudyFiles(
  path = path,
  outputDir = outputDir
)
unlink(outputDir, recursive = TRUE)
}
}
