% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipStudyFiles.R
\name{unZipStudyFiles}
\alias{unZipStudyFiles}
\title{`unZipStudyFiles()` extract results for importing}
\usage{
unZipStudyFiles(path, pattern, negate = FALSE, recursive = FALSE, outputDir)
}
\arguments{
\item{path}{The path to the directory where the .zip files are located}

\item{pattern}{A character string to filter the files in the location provided, default NULL}

\item{negate}{If TRUE it will filter the opposite specified in pattern, default FALSE}

\item{recursive}{If TRUE it will search recursively for .zip files in the path}

\item{outputDir}{Character string. The path to the main output folder}
}
\value{
A message stating on where the result files were extracted
}
\description{
`unZipStudyFiles()` extract results for importing
}
