% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateColumnValues.R
\name{updateColumnValues}
\alias{updateColumnValues}
\title{Update values in a summarised result}
\usage{
updateColumnValues(summarised_result, names_map, variable)
}
\arguments{
\item{summarised_result}{The summarised result with values to update.}

\item{names_map}{A named vector containing the mapping between old and new names.}

\item{variable}{The column name(s) containing the values to be updated in the
summarised result object.}
}
\value{
The summarised result itself with updated values.
}
\description{
Updates the character values of specific column(s) in a summarised result object
(e.g., changing cohort names to a more polished version for Shiny app labels)
according to a given name mapping.
}
