% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipStudyFiles.R
\name{zipStudyFiles}
\alias{zipStudyFiles}
\title{\code{zipStudyFiles()} compress study results}
\usage{
zipStudyFiles(outputDir, resultsDirName, dbname)
}
\arguments{
\item{outputDir}{Character path to the base output directory.}

\item{resultsDirName}{Character name or path to the results subfolder.}

\item{dbname}{Character database identifier.}
}
\value{
A message stating the location of the compressed results

Character file name of the generated zip archive.
}
\description{
\code{zipStudyFiles()} compress study results
}
\examples{
outputDir <- file.path(
   tempdir(),
   "examples"
   )
dbname <- "IPCI"
directories <- createResultsDir(
  outputDir,
  dbname = dbname
)
zipStudyFiles(
  resultsDirName = directories$resultsDir,
  outputDir = directories$outputDir,
  dbname = dbname
)
unlink(outputDir, recursive = TRUE)
}
