% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipStudyFiles.R
\name{zipStudyFiles}
\alias{zipStudyFiles}
\title{`zipStudyFiles()` compress study results}
\usage{
zipStudyFiles(resultsDirName, outputDir, dbname)
}
\arguments{
\item{resultsDirName}{Location of the results. Character.}

\item{outputDir}{The directory where the results folder is located. Character}

\item{dbname}{The database name. Character}
}
\value{
A message stating the location of the compressed results
}
\description{
`zipStudyFiles()` compress study results
}
